+++ ws/snake/IXSnakeServer.cpp
@@ -58,25 +58,20 @@ namespace snake
                 auto state = std::dynamic_pointer_cast<SnakeConnectionState>(connectionState);
 
                 webSocket->setOnMessageCallback(
-                    [this, webSocket, state](ix::WebSocketMessageType messageType,
-                       const std::string& str,
-                       size_t wireSize,
-                       const ix::WebSocketErrorInfo& error,
-                       const ix::WebSocketOpenInfo& openInfo,
-                       const ix::WebSocketCloseInfo& closeInfo)
+                    [this, webSocket, state](const ix::WebSocketMessagePtr& msg)
                     {
-                        if (messageType == ix::WebSocketMessageType::Open)
+                        if (msg->type == ix::WebSocketMessageType::Open)
                         {
                             std::cerr << "New connection" << std::endl;
                             std::cerr << "id: " << state->getId() << std::endl;
-                            std::cerr << "Uri: " << openInfo.uri << std::endl;
+                            std::cerr << "Uri: " << msg->openInfo.uri << std::endl;
                             std::cerr << "Headers:" << std::endl;
-                            for (auto it : openInfo.headers)
+                            for (auto it : msg->openInfo.headers)
                             {
                                 std::cerr << it.first << ": " << it.second << std::endl;
                             }
 
-                            std::string appkey = parseAppKey(openInfo.uri);
+                            std::string appkey = parseAppKey(msg->openInfo.uri);
                             state->setAppkey(appkey);
 
                             // Connect to redis first
@@ -86,29 +81,29 @@ namespace snake
                                 std::cerr << "Cannot connect to redis host" << std::endl;
                             }
                         }
-                        else if (messageType == ix::WebSocketMessageType::Close)
+                        else if (msg->type == ix::WebSocketMessageType::Close)
                         {
                             std::cerr << "Closed connection"
-                                      << " code " << closeInfo.code
-                                      << " reason " << closeInfo.reason << std::endl;
+                                      << " code " << msg->closeInfo.code
+                                      << " reason " << msg->closeInfo.reason << std::endl;
                         }
-                        else if (messageType == ix::WebSocketMessageType::Error)
+                        else if (msg->type == ix::WebSocketMessageType::Error)
                         {
                             std::stringstream ss;
-                            ss << "Connection error: " << error.reason      << std::endl;
-                            ss << "#retries: "         << error.retries     << std::endl;
-                            ss << "Wait time(ms): "    << error.wait_time   << std::endl;
-                            ss << "HTTP Status: "      << error.http_status << std::endl;
+                            ss << "Connection error: " << msg->errorInfo.reason      << std::endl;
+                            ss << "#retries: "         << msg->errorInfo.retries     << std::endl;
+                            ss << "Wait time(ms): "    << msg->errorInfo.wait_time   << std::endl;
+                            ss << "HTTP Status: "      << msg->errorInfo.http_status << std::endl;
                             std::cerr << ss.str();
                         }
-                        else if (messageType == ix::WebSocketMessageType::Fragment)
+                        else if (msg->type == ix::WebSocketMessageType::Fragment)
                         {
                             std::cerr << "Received message fragment" << std::endl;
                         }
-                        else if (messageType == ix::WebSocketMessageType::Message)
+                        else if (msg->type == ix::WebSocketMessageType::Message)
                         {
-                            std::cerr << "Received " << wireSize << " bytes" << std::endl;
-                            processCobraMessage(state, webSocket, _appConfig, str);
+                            std::cerr << "Received " << msg->wireSize << " bytes" << std::endl;
+                            processCobraMessage(state, webSocket, _appConfig, msg->str);
                         }
                     }
                 );