+++ ws/ws_transfer.cpp
@@ -21,52 +21,47 @@ namespace ix
                       std::shared_ptr<ConnectionState> connectionState)
             {
                 webSocket->setOnMessageCallback(
-                    [webSocket, connectionState, &server](ix::WebSocketMessageType messageType,
-                       const std::string& str,
-                       size_t wireSize,
-                       const ix::WebSocketErrorInfo& error,
-                       const ix::WebSocketOpenInfo& openInfo,
-                       const ix::WebSocketCloseInfo& closeInfo)
+                    [webSocket, connectionState, &server](const WebSocketMessagePtr& msg)
                     {
-                        if (messageType == ix::WebSocketMessageType::Open)
+                        if (msg->type == ix::WebSocketMessageType::Open)
                         {
                             std::cerr << "New connection" << std::endl;
                             std::cerr << "id: " << connectionState->getId() << std::endl;
-                            std::cerr << "Uri: " << openInfo.uri << std::endl;
+                            std::cerr << "Uri: " << msg->openInfo.uri << std::endl;
                             std::cerr << "Headers:" << std::endl;
-                            for (auto it : openInfo.headers)
+                            for (auto it : msg->openInfo.headers)
                             {
                                 std::cerr << it.first << ": " << it.second << std::endl;
                             }
                         }
-                        else if (messageType == ix::WebSocketMessageType::Close)
+                        else if (msg->type == ix::WebSocketMessageType::Close)
                         {
                             std::cerr << "Closed connection"
-                                      << " code " << closeInfo.code
-                                      << " reason " << closeInfo.reason << std::endl;
+                                      << " code " << msg->closeInfo.code
+                                      << " reason " << msg->closeInfo.reason << std::endl;
                         }
-                        else if (messageType == ix::WebSocketMessageType::Error)
+                        else if (msg->type == ix::WebSocketMessageType::Error)
                         {
                             std::stringstream ss;
-                            ss << "Connection error: " << error.reason      << std::endl;
-                            ss << "#retries: "         << error.retries     << std::endl;
-                            ss << "Wait time(ms): "    << error.wait_time   << std::endl;
-                            ss << "HTTP Status: "      << error.http_status << std::endl;
+                            ss << "Connection error: " << msg->errorInfo.reason      << std::endl;
+                            ss << "#retries: "         << msg->errorInfo.retries     << std::endl;
+                            ss << "Wait time(ms): "    << msg->errorInfo.wait_time   << std::endl;
+                            ss << "HTTP Status: "      << msg->errorInfo.http_status << std::endl;
                             std::cerr << ss.str();
                         }
-                        else if (messageType == ix::WebSocketMessageType::Fragment)
+                        else if (msg->type == ix::WebSocketMessageType::Fragment)
                         {
                             std::cerr << "Received message fragment "
                                       << std::endl;
                         }
-                        else if (messageType == ix::WebSocketMessageType::Message)
+                        else if (msg->type == ix::WebSocketMessageType::Message)
                         {
-                            std::cerr << "Received " << wireSize << " bytes" << std::endl;
+                            std::cerr << "Received " << msg->wireSize << " bytes" << std::endl;
                             for (auto&& client : server.getClients())
                             {
                                 if (client != webSocket)
                                 {
-                                    client->send(str,
+                                    client->send(msg->str,
                                                  [](int current, int total) -> bool
                                     {
                                         std::cerr << "ws_transfer: Step " << current